This commit fixes a problem when calculating the position of beginning of defun for preselect'ed helm-gtags-parse-file. Previously, it uses line-number-at-pos. This works only when the cursor is on the same line as defun. It's more useful if preselect is available when the cursor is not on the same line as defun.
To achieve this, I tried helm-gtags--current-funcion-bound, but I faced another problem. The helm-gtags--current-funcion-bound uses beginning-of-defun to find the position of defun, however, return value of beginning-of-defun is not necesarrily same as the value retrieved from global command. For example, line numbers will mismatch between beginning-of-defun and global.

1: void
2: main()
3: {
4:   puts("Hello world!");
5: }

In this case, beginning-of-defun moves to line 1 while global returns line 2 on the other hand. Due to this mismatch, simple regex based comparison doesn't work well. The simplest way to solve this problem is to use global to construct preselect regex pattern. In this patch, I introduced several functions to do it.